Added auth0 service
In this commit I added the dependencies to use the auth0 service. The
API keys to be used will be sent to the team to actually test out the
authentication system. This commit is just to configure the dependency
globally, so that later commits can use the authentication system after
the page is added.